feat: Update dependencies and integrate site visit link
Upgrades project dependencies to newer versions and introduces a direct link to the public-facing website from the admin dashboard. This ensures compatibility with the latest tools and provides convenient access for users to view the live site.

Additionally, styling and script for the public site have been partially updated and included.